Create Index file for a logbook
This format was completely stolen from the BMX github repo, but I liked it and I hope you all will too. 

If we like the format of making postings to discuss results/figures, this will serve as an index where we link to the markdown files that discuss the postings, which are best viewed in an internet browser. Then images/figures linked in the markdown file will appear in an article format.